back/requirements.txt
fastapi==0.115.14
uvicorn[standard]
pyjwt==2.10.1
passlib==1.7.4
